refactor: Use the Navigation Arch Component to navigate between screens - Create a Navigation Graph - Create a MainActivity to be the navigation host - Replaced activities with fragments to be hosted on the MainActivity
