+++ app/build.gradle
@@ -33,6 +33,13 @@ dependencies {
 
     implementation "android.arch.lifecycle:extensions:1.1.1"
 
+    implementation "android.arch.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "android.arch.navigation:navigation-ui-ktx:$nav_version"
+
+    // optional - Test helpers
+    implementation 'com.android.support:support-v4:28.0.0-alpha1'
+    androidTestImplementation "android.arch.navigation:navigation-testing:$nav_version"
+
     implementation "com.shopify:livedata-ktx:1.0.1"
 
     implementation 'com.squareup.retrofit2:retrofit:2.4.0'
