+++ app/src/main/java/com/myshopify/shopicruit/shopifyinternchallenge/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.myshopify.shopicruit.shopifyinternchallenge
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import androidx.navigation.findNavController
+import kotlinx.android.synthetic.main.activity_main.*
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setSupportActionBar(toolbar)
+
+    }
+
+    override fun onSupportNavigateUp() = findNavController(R.id.nav_host_fragment).navigateUp()
+
+    companion object {
+        const val PRODUCT_TAG_KEY = "com.shopify.shopicruit.product.tag"
+    }
+}
