+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/tagsFragment">
+
+    <fragment
+        android:id="@+id/tagsFragment"
+        android:name="com.myshopify.shopicruit.shopifyinternchallenge.tagslist.TagsFragment"
+        android:label="TagsFragment"
+        tools:layout="@layout/content_simple_list">
+        <action
+            android:id="@+id/action_tagsFragment_to_productsFragment"
+            app:destination="@id/productsFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/productsFragment"
+        android:name="com.myshopify.shopicruit.shopifyinternchallenge.productslist.ProductsFragment"
+        android:label="ProductsFragment"
+        tools:layout="@layout/content_simple_list"/>
+</navigation>
