+++ build.gradle
@@ -2,6 +2,8 @@
 
 buildscript {
     ext.kotlin_version = '1.2.60'
+    ext.nav_version = "1.0.0-alpha05"
+
     repositories {
         google()
         jcenter()
